requirements.txt
annotated-types==0.5.0
anyio==3.7.1
bcrypt==4.0.1
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
deep-translator==1.11.4
ecdsa==0.18.0
fastapi==0.103.1
filelock==3.12.3
fsspec==2023.9.0
greenlet==2.0.2
h11==0.14.0
huggingface-hub==0.16.4
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.1
numpy==1.25.2
packaging==23.1
pandas==2.1.0
pyasn1==0.5.0
pydantic==2.3.0
pydantic-settings==2.0.3
pydantic_core==2.6.3
PyMySQL==1.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
rsa==4.9
safetensors==0.3.3
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.20
starlette==0.27.0
sympy==1.12
tokenizers==0.13.3
torch==2.0.1
tqdm==4.66.1
transformers==4.32.1
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
uvicorn==0.23.2